Added examples

+++ examples/custom_key_transformation.cpp
@@ -8,16 +8,16 @@
 // positive ones. So, this is kind of an "incorrect" transformation,
 // but it shows the intricacies of comparing at bit-level.
 struct custom_transform {
-    int64_t operator()(const int64_t &key) const noexcept {
+    int operator()(const int &key) const noexcept {
         return art::is_big_endian() ? key : art::byte_swap(key);
     }
 };
 
 int main() {
-    art::radix_map<int64_t, int64_t, custom_transform> map;
-    map.insert(std::pair<int64_t, int64_t>(-2, 6));
-    map.insert(std::pair<int64_t, int64_t>(0, 5));
-    map.insert(std::pair<int64_t, int64_t>(7, 25));
+    art::radix_map<int, int, custom_transform> map;
+    map.insert(std::pair<int, int>(-2, 6));
+    map.insert(std::pair<int, int>(0, 5));
+    map.insert(std::pair<int, int>(7, 25));
 
     for (auto &e : map)
         std::cout << e.first << " -> " << e.second << std::endl;
